+++ backend/common/serializers.py
@@ -8,6 +8,7 @@ from .models import (
     PoolingTemplate,
     RunStatisticsTemplate,
     SequencesStatisticsTemplate,
+    InvoicingTemplate,
 )
 
 from .models import CostUnit, Duty, User
@@ -76,4 +77,10 @@ class RunStatisticsTemplateSerializer(ModelSerializer):
 class SequencesStatisticsTemplateSerializer(ModelSerializer):
     class Meta:
         model = SequencesStatisticsTemplate
+        fields = ["id", "name", "file", "uploaded_at"]
+
+
+class InvoicingTemplateSerializer(ModelSerializer):
+    class Meta:
+        model = InvoicingTemplate
         fields = ["id", "name", "file", "uploaded_at"]